feat(wechat-public-account-push): 新增天气字段
- 优化仓库目录结构，主脚本代码更清爽了！

- 增加更多天气预设的字段, 由用户自由选择想要展示的字段
